# sandbox/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Docker 샌드박스 컨테이너 관리를 위한 확장 가능한 API 서버입니다.",
        "title": "Sandbox API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/sandbox/run": {
            "post": {
                "description": "요청된 시나리오와 사용자 정보에 매칭되는 샌드박스 컨테이너를 구동합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sandbox"
                ],
                "summary": "샌드박스 컨테이너 실행",
                "parameters": [
                    {
                        "description": "샌드박스 실행 파라미터",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SandboxRunRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SandboxRunResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.SandboxRunRequest": {
            "type": "object",
            "required": [
                "scenario",
                "user"
            ],
            "properties": {
                "scenario": {
                    "description": "실행할 시나리오 이름 (필수)",
                    "type": "string",
                    "example": "python-basic"
                },
                "user": {
                    "description": "요청 사용자 식별자 (필수)",
                    "type": "string",
                    "example": "test-user"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "에러 상세 메시지",
                    "type": "string",
                    "example": "invalid request parameters"
                }
            }
        },
        "response.SandboxRunResponse": {
            "type": "object",
            "properties": {
                "container_id": {
                    "description": "생성된 컨테이너 식별자 ID",
                    "type": "string",
                    "example": "sandbox-test-user"
                },
                "status": {
                    "description": "현재 컨테이너 상태",
                    "type": "string",
                    "example": "running"
                }
            }
        }
    }
}